tests/requirements.txt
Django>=1.5
elasticsearch>=1.5.0
urllib3>=1.10.4
psycopg2
ipdb
django-haystack>=2.3.1
django-modeltranslation==0.9.1
